Issue 28 - Show error if users try to delete related models

+++ routes/PHDTags.ts
@@ -147,11 +147,17 @@ router.delete("/:id", async (req, res) => {
 
   const tag = await prisma.pHDTag.findUnique({
     where: { id },
+    include: { assignment: true },
   });
   if (!tag)
     return res
       .status(404)
       .send({ message: "The PHD tag with the given ID was not found." });
+
+  if (tag.assignment.length)
+    return res
+      .status(500)
+      .send({ message: "The tag has one or more assignments" });
 
   await prisma.pHDTag.delete({ where: { id } });
