+++ routes/units.ts
@@ -129,11 +129,18 @@ router.delete("/:id", async (req, res) => {
 
   const unit = await prisma.unit.findUnique({
     where: { id },
+    include: { PHDTag: true },
   });
   if (!unit)
     return res
       .status(404)
       .send({ message: "The unit with the given ID was not found." });
+
+  if (unit.PHDTag.length)
+    return res.status(500).send({
+      message:
+        "There are assignments to the units. Please delete according assigned tags",
+    });
 
   await prisma.unit.delete({ where: { id } });
